Masukkan modifan ke form aslinya
