+++ src/simrskhanza/frmUtama.java
@@ -665,7 +665,6 @@ import keuangan.KeuanganPiutangBelumLunas;
 import keuangan.KeuanganPiutangLainLain;
 import keuangan.KeuanganPiutangObatBelumLunas;
 import keuangan.KeuanganRVPBPJS;
-import keuangan.KeuanganRVPBPJSSMC;
 import keuangan.KeuanganRekapJmDokter;
 import keuangan.KeuanganRekapPengajuanBiaya;
 import keuangan.KeuanganRekapPoliAnak;
@@ -754,9 +753,9 @@ import laporan.DlgPenyakit;
 import laporan.DlgPenyakitPd3i;
 import laporan.DlgPenyakitRanapPerCaraBayar;
 import laporan.DlgPenyiapanRM;
-import laporan.DlgRL13KetersediaanTempatTidur;
 import laporan.DlgPerujukLabPerTahun;
 import laporan.DlgPerujukRadiologiPerTahun;
+import laporan.DlgRL13KetersediaanTempatTidur;
 import laporan.DlgRL4A;
 import laporan.DlgRL4ASebab;
 import laporan.DlgRL4B;
@@ -798,9 +797,9 @@ import permintaan.DlgBookingRegistrasi;
 import permintaan.DlgCariPermintaanLab;
 import permintaan.DlgCariPermintaanLabMB;
 import permintaan.DlgCariPermintaanLabPA;
+import permintaan.DlgCariPermintaanLayananProgramKFR;
 import permintaan.DlgCariPermintaanRadiologi;
 import permintaan.DlgPermintaanKonsultasiMedik;
-import permintaan.DlgCariPermintaanLayananProgramKFR;
 import permintaan.DlgPermintaanPelayananInformasiObat;
 import permintaan.DlgPermintaanRanap;
 import perpustakaan.PerpustakaanAnggota;
@@ -987,9 +986,10 @@ import rekammedis.RMSkriningInstrumenSDQ;
 import rekammedis.RMSkriningKankerKolorektal;
 import rekammedis.RMSkriningKekerasanPadaPerempuan;
 import rekammedis.RMSkriningKesehatanGigiMulutBalita;
-import rekammedis.RMSkriningKesehatanGigiMulutRemaja;
 import rekammedis.RMSkriningKesehatanGigiMulutDewasa;
 import rekammedis.RMSkriningKesehatanGigiMulutLansia;
+import rekammedis.RMSkriningKesehatanGigiMulutRemaja;
+import rekammedis.RMSkriningKesehatanPenglihatan;
 import rekammedis.RMSkriningMPP;
 import rekammedis.RMSkriningMPPFormA;
 import rekammedis.RMSkriningMPPFormB;
@@ -998,7 +998,6 @@ import rekammedis.RMSkriningNutrisiAnak;
 import rekammedis.RMSkriningNutrisiDewasa;
 import rekammedis.RMSkriningNutrisiLansia;
 import rekammedis.RMSkriningObesitas;
-import rekammedis.RMSkriningKesehatanPenglihatan;
 import rekammedis.RMSkriningPUMA;
 import rekammedis.RMSkriningRisikoKankerParu;
 import rekammedis.RMSkriningRisikoKankerPayudara;
@@ -17472,7 +17471,7 @@ private void MnGantiPasswordBtnLogActionPerformed(java.awt.event.ActionEvent evt
     private void btnRVPPiutangBPJSActionPerformed(java.awt.event.ActionEvent evt) { 
         isTutup();
         this.setCursor(Cursor.getPredefinedCursor(Cursor.WAIT_CURSOR));
-        KeuanganRVPBPJSSMC rbpaketbhp=new KeuanganRVPBPJSSMC(this,false);
+        KeuanganRVPBPJS rbpaketbhp=new KeuanganRVPBPJS(this,false);
         rbpaketbhp.isCek();
         rbpaketbhp.setSize(PanelUtama.getWidth(),PanelUtama.getHeight());
         rbpaketbhp.setLocationRelativeTo(PanelUtama);
